pyproject.toml
[tool.poetry]
name = "monolithicflask"
version = "0.1.0"
description = "FastAPI Monolithic Web Application"
authors = ["Daniil Protasenia <danikprotaseney@gmail.com>"]
readme = "README.md"

# System
[tool.poetry.dependencies]
python = "^3.12"

# Poetry core
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Dev
[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

# App
[tool.poetry.group.app]
optional = true

[tool.poetry.group.app.dependencies]
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
jinja2 = "^3.1.3"
python-multipart = "^0.0.7"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
redis = "^5.0.1"
aioredis = "^2.0.1"
boto3 = "^1.34.38"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"

# Test
[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
coverage = "^7.4.1"

# Configs
## Linters
[tool.flake8]
exclude = [".git", "__pycache__", "env", "venv"]
max-line-length = 120
max-complexity = 8

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.black]
line-length = 120
